Move projectInputVerification in displayProjects to logic.js
